// zebra-ui/src/interfaces/api/corporate-domain.ts
import { CorporateDomainStatus } from "@/common/enums";

export interface CorporateDomain {
  id: number;
  domainName: string;
  status: CorporateDomainStatus;
  websiteTitle: string;
  createdAt: string;
  updatedAt: string;
}

export interface GetCorporateDomainsResponse {
  content: CorporateDomain[];
  pageable: {
    pageNumber: number;
    pageSize: number;
    sort: {
      empty: boolean;
      sorted: boolean;
      unsorted: boolean;
    };
    offset: number;
    paged: boolean;
    unpaged: boolean;
  };
  last: boolean;
  totalPages: number;
  totalElements: number;
  size: number;
  number: number;
  sort: {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
  };
  first: boolean;
  numberOfElements: number;
  empty: boolean;
}